clients/python/agentic-sandbox-client/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic_sandbox"
version = "0.1.0"
description = "A client library to interact with the Agentic Sandbox on Kubernetes."
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
dependencies = [
    "kubernetes",
    "requests",
]

[project.urls]
"Homepage" = "https://github.com/kubernetes-sigs/agent-sandbox"
"Bug Tracker" = "https://github.com/kubernetes-sigs/agent-sandbox/issues"

[tool.setuptools]
# This line tells setuptools to only include the 'agentic_sandbox'
# package and to ignore other directories like 'sandbox_router'.
packages = ["agentic_sandbox"]

[project.optional-dependencies]
langchain = [
    "langchain",
]
adk = [
    "google-adk",
]
crewai = [
    "crewai[tools]",
]
test = [
    "agentic_sandbox[langchain]",
    "agentic_sandbox[adk]",
    "agentic_sandbox[crewai]",
    "pytest",
    "pytest-xdist",
]
[tool.pytest]
pythonpath = [
    "."
]
